[new] frontend

+++ backend/src/app.py
@@ -10,6 +10,8 @@ from Domain.user import*
 from Domain.dealership import*
 from Domain.car import*
 from flask_cors import CORS
+from jwt import TokenStorage
+from OpenSSL import SSL
 
 import argparse
 
@@ -20,6 +22,7 @@ fabrique = None
 user_repository = None
 dealership_respository = None
 car_repository = None
+token_storage = TokenStorage()
 
 api_doc(app, config_path='../config/swagger.yaml', url_prefix='/api/v1', title='API doc')
 
@@ -66,6 +69,7 @@ def register_user():
         return send_error(400)
 
     new_user_domain = User(
+        None,
         req['username'],
         req['password'],
         req['role']
@@ -75,25 +79,42 @@ def register_user():
     for user in users:
         if user.username == new_user_domain.username:
             return send_error(400)
-    
+
     user_repository.add_new_user(new_user_domain)
 
-    return send_ok()
+    users = user_repository.get_all_users()
+
+    for user in users:
+        if user.username == new_user_domain.username:
+            token_string = token_storage.create_token()
+            response = {
+                'id': user.id,
+                'role': new_user_domain.role,
+                'token': token_string
+            }
+            return jsonify(response), 200
+
+    return send_error(400)
 
 @app.route("/api/v1/users/login", methods=['POST'])
 def login_user():
     req = request.json
     if not req or not 'username' in req or not 'password' in req:
         return send_error(400)
-
     users = user_repository.get_all_users()
     username = req['username']
     password = req['password']
-
     for user in users:
         if user.username == username:
             if user.password == password:
-                return send_ok()
+                print(users)
+                token_string = token_storage.create_token()
+                response = {
+                    'id': user.id,
+                    'role': user.role,
+                    'token': token_string
+                }
+                return jsonify(response), 200
             else:
                 return send_error(400)
 
@@ -101,6 +122,10 @@ def login_user():
 
 @app.route("/api/v1/dealerships", methods=['GET'])
 def get_dealerships():
+    token_string = request.headers['Authorization']
+    if not token_storage.token_exists(token_string):
+        return send_error(401)
+
     dealerships = dealership_respository.get_dealerships()
     DTO_dealerships = []
     converter = DealershipDomainToDTOConverter()
@@ -112,6 +137,9 @@ def get_dealerships():
 
 @app.route("/api/v1/dealerships/dealership", methods=['POST'])
 def create_dealership():
+    token_string = request.headers['Authorization']
+    if not token_storage.token_exists(token_string):
+        return send_error(401)
     req = request.json
     if not req or not 'name' in req or not 'description' in req or not 'owner_id' in req:
         return send_error(400)
@@ -128,7 +156,7 @@ def create_dealership():
     for dealer in dealerships:
         if dealer.name == new_dealership_domain.name:
             return send_error(400)
-
+    print(new_dealership_domain)
     try:
         dealership_respository.create_dealership(new_dealership_domain)
     except:
@@ -138,6 +166,9 @@ def create_dealership():
 
 @app.route("/api/v1/dealerships/<int:id>", methods=['DELETE'])
 def delete_dealership(id):
+    token_string = request.headers['Authorization']
+    if not token_storage.token_exists(token_string):
+        return send_error(401)
     try:
         dealership_respository.remove_dealership(id)
     except:
@@ -147,6 +178,9 @@ def delete_dealership(id):
 
 @app.route("/api/v1/cars/<int:dealership_id>", methods=['GET'])
 def get_cars_for_dealership(dealership_id):
+    token_string = request.headers['Authorization']
+    if not token_storage.token_exists(token_string):
+        return send_error(401)
     cars = car_repository.get_cars(dealership_id)
 
     converter = CarDomainToDTOConverter()
@@ -159,6 +193,9 @@ def get_cars_for_dealership(dealership_id):
 
 @app.route("/api/v1/cars/car/<int:id>", methods=['DELETE'])
 def delete_car(id):
+    token_string = request.headers['Authorization']
+    if not token_storage.token_exists(token_string):
+        return send_error(401)
     try:
         car_repository.delete_car(id)
     except:
@@ -168,6 +205,9 @@ def delete_car(id):
 
 @app.route("/api/v1/cars/car", methods=['POST'])
 def create_car():
+    token_string = request.headers['Authorization']
+    if not token_storage.token_exists(token_string):
+        return send_error(401)
     req = request.json
     if not req\
         or not 'model' in req\
@@ -192,6 +232,9 @@ def create_car():
 
 @app.route("/api/v1/cars/car/availabilty", methods=['PATCH'])
 def change_car_availability():
+    token_string = request.headers['Authorization']
+    if not token_storage.token_exists(token_string):
+        return send_error(401)
     req = request.json
     if not req or not 'is_available' in req or not 'car_id' in req:
         return send_error(400)
@@ -223,4 +266,4 @@ if __name__ == "__main__":
 
     cors = CORS(app)
     app.logger.disabled = True
-    app.run(debug=False,port=int(args.back_port))
+    app.run(debug=False,port=int(args.back_port),ssl_context=('localhost.crt', 'localhost.key'))
